+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+* Remove the part of the body from the error message for SourceWrapperNotFoundError and do some cleanup
+  Use more explanatory error message. 
+
 # 18.5.2
 
 * Fix a mistake in 18.5.1